server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.1.2",
    "fastapi>=0.123.4",
    "httpx>=0.28.1",
    "motor>=3.7.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]==1.7.4",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-playwright>=0.7.2",
]
